Update playwright data structure
As we migrate to Playwright we want to improve our scenarios and how we manage and access the data.

This change separates the existing licence scenario into unregistered and registered licence scenarios.

Where possible, we will endeavour to have related data colocated inside its relevant data file. This means, for example, the primary user data file will provide all the data required to construct a primary user for a licence.

Spec files now destructure the scenario data using array patterns, for example:

  const {
    licences: [licence]
  } = scenario

instead of indexing directly (scenario.licences[0]), making it clearer which arrays a test depends on.

- tests/support/data/company.js is a new file (previously empty), extracted from the old licence.js.
- tests/support/data/users.js now exports a named primaryUser function alongside the existing users object (was a default export before); fixtures.js was updated for that import change too.
