+++ tests/internal/licence/search.spec.js
@@ -2,7 +2,10 @@ import scenarioData from '../../support/scenarios/licence.js'
 import { test, expect } from '../../support/fixtures.js'
 
 const scenario = scenarioData()
-const { licenceRef } = scenario.licences[0]
+
+const {
+  licences: [licence]
+} = scenario
 
 test.describe('Search for a licence (internal)', () => {
   test.beforeAll(async ({ setup }) => {
@@ -15,22 +18,22 @@ test.describe('Search for a licence (internal)', () => {
 
   test('can find a licence by exact licence reference', async ({ page }) => {
     await page.goto('/')
-    await page.locator('#query').fill(licenceRef)
+    await page.locator('#query').fill(licence.licenceRef)
     await page.locator('#search-button').click()
-    await expect(page.locator('.searchresult-row')).toContainText(licenceRef)
+    await expect(page.locator('.searchresult-row')).toContainText(licence.licenceRef)
   })
 
   test('can find a licence by lowercase licence reference', async ({ page }) => {
     await page.goto('/')
-    await page.locator('#query').fill(licenceRef.toLowerCase())
+    await page.locator('#query').fill(licence.licenceRef.toLowerCase())
     await page.locator('#search-button').click()
-    await expect(page.locator('.searchresult-row')).toContainText(licenceRef)
+    await expect(page.locator('.searchresult-row')).toContainText(licence.licenceRef)
   })
 
   test('can find a licence by partial licence reference', async ({ page }) => {
     await page.goto('/')
-    await page.locator('#query').fill(licenceRef.slice(0, -1))
+    await page.locator('#query').fill(licence.licenceRef.slice(0, -1))
     await page.locator('#search-button').click()
-    await expect(page.locator('.searchresult-row')).toContainText(licenceRef)
+    await expect(page.locator('.searchresult-row')).toContainText(licence.licenceRef)
   })
 })
