+++ tests/internal/licence/unregister.spec.js
@@ -1,10 +1,13 @@
-import scenarioData from '../../support/scenarios/licence.js'
+import scenarioData from '../../support/scenarios/registered-licence.js'
 import { test, expect } from '../../support/fixtures.js'
 
 const scenario = scenarioData()
-const { licenceRef } = scenario.licences[0]
-const { name: companyName } = scenario.companies[0]
-const { username: userEmail } = scenario.users[0]
+
+const {
+  companies: [company],
+  licences: [licence],
+  users: [user]
+} = scenario
 
 test.describe('Unregister a licence (internal)', () => {
   test.beforeAll(async ({ setup }) => {
@@ -19,9 +22,9 @@ test.describe('Unregister a licence (internal)', () => {
     await page.goto('/')
 
     // Search for the user and then select them
-    await page.locator('#query').fill(userEmail)
+    await page.locator('#query').fill(user.username)
     await page.locator('#search-button').click()
-    await expect(page.locator('.searchresult-row')).toContainText(userEmail)
+    await expect(page.locator('.searchresult-row')).toContainText(user.username)
     await page.locator('.searchresult-link').click()
 
     // Select the external user's licences page
@@ -31,13 +34,15 @@ test.describe('Unregister a licence (internal)', () => {
     await page.locator('.govuk-button').filter({ hasText: 'Unregister licence' }).click()
 
     // Select the licence to be unregistered
-    await expect(page.locator('.govuk-label')).toContainText(licenceRef)
-    await expect(page.locator('#licences-item-hint')).toContainText(companyName)
+    await expect(page.locator('.govuk-label')).toContainText(licence.licenceRef)
+    await expect(page.locator('#licences-item-hint')).toContainText(company.name)
     await page.locator('[name="licences"]').click()
     await page.locator('.govuk-button').filter({ hasText: 'Continue' }).click()
 
     // Confirm licences to unregister
-    await expect(page.locator('#main-content > dl > div > dd.govuk-summary-list__value > p')).toContainText(licenceRef)
+    await expect(page.locator('#main-content > dl > div > dd.govuk-summary-list__value > p')).toContainText(
+      licence.licenceRef
+    )
     await page.locator('button.govuk-button').filter({ hasText: 'Confirm' }).click()
 
     // Confirm notification and licence is no longer shown
@@ -46,9 +51,9 @@ test.describe('Unregister a licence (internal)', () => {
 
     // Confirm the licence is now shown as unregistered
     await page.locator('#nav-search').click()
-    await page.locator('#query').fill(licenceRef)
+    await page.locator('#query').fill(licence.licenceRef)
     await page.locator('#search-button').click()
-    await expect(page.locator('.searchresult-row')).toContainText(licenceRef)
+    await expect(page.locator('.searchresult-row')).toContainText(licence.licenceRef)
     await page.locator('.searchresult-link').click()
 
     await expect(page.locator('.govuk-caption-l')).toContainText('Unregistered licence')
