+++ tests/support/fixtures.js
@@ -1,7 +1,7 @@
 import { test as base } from '@playwright/test'
 import { readFileSync } from 'fs'
 
-import users from './data/users.js'
+import { users } from './data/users.js'
 
 const environment = process.env.TEST_ENV ?? 'local'
 const envConfig = JSON.parse(readFileSync(`./environments/${environment}.json`, 'utf8'))
