+++ tests/support/scenarios/licence.js
@@ -1,8 +1,16 @@
 import licence from '../data/licence.js'
+import company from '../data/company.js'
 
-export const title = 'Licence, with Primary user'
-export const description = 'A registered licence'
+export const title = 'Licence (unregistered)'
+export const description = 'A licence, with a company and licence holder'
 
 export default function () {
-  return licence()
+  const licenceRef = 'AT/TE/ST/01/01'
+
+  const companies = company('Big Farm Co Ltd')
+
+  return {
+    ...companies,
+    ...licence(licenceRef, companies)
+  }
 }
